+++ lib/src/issue_table.dart
@@ -161,13 +161,6 @@ TextStyle? rowStyle(Issue issue) {
   return null;
 }
 
-int compareDates(DateTime? a, DateTime? b) {
-  if (a == b) return 0;
-  if (a == null) return -1;
-  if (b == null) return 1;
-  return a.compareTo(b);
-}
-
 class LabelWidget extends StatelessWidget {
   final IssueLabel label;
 
@@ -200,9 +193,6 @@ class LabelWidget extends StatelessWidget {
     );
   }
 }
-
-bool isLightColor(Color color) =>
-    ThemeData.estimateBrightnessForColor(color) == Brightness.light;
 
 ValidationResult? oldPrValidator(Issue issue) {
   final createdAt = issue.createdAt;
